requirements.txt
alembic==1.4.1
beautifulsoup4==4.8.2
blessings==1.7
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
curtsies==0.3.1
Flask==1.1.1
flask-marshmallow==0.11.0
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
ForgeryPy==0.1
greenlet==0.4.15
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.1
mysql-connector==2.2.9
Pygments==2.5.2
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.13
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.0
